fea-38 refactor the ArticleImage component to TS, and refactor the corresponding story and ensure use of Controls vs knobs.
