+++ src/index.js
@@ -5,7 +5,7 @@ import AccordionControl from './components/AccordionControl/AccordionControl';
 import AccordionRefinementList from './components/Algolia/shared/AccordionRefinementList';
 import ActionSummaryItem from './components/ActionSummaryItem/ActionSummaryItem';
 import ArticleFigcaption from './components/Articles/shared/ArticleFigcaption';
-import ArticleImage from './components/Articles/ArticleImage';
+import ArticleImage from './components/Articles/ArticleImage/ArticleImage';
 import ArticlePhotoCollection from './components/Articles/ArticlePhotoCollection';
 import ArticleTextBlock from './components/Articles/ArticleTextBlock';
 import AtkMarketingHat from './components/MarketingHat';
